Warning fix

+++ OpenGLHelper.h
@@ -20,7 +20,7 @@ public:
 	~OpenGLHelper();
 
 	void load_texture(unsigned char* data, int width, int height, int nrComponents, GLuint textureID);
-	GLuint get_output_texture_id() { return output_texture_id; };	// output texture id
+	size_t get_output_texture_id() { return output_texture_id; };	// output texture id
 	size_t get_texture_id_at_slot(uint8_t slot);	// returns the opengl-generated texture id for this tex slot
 
 	// TODO: Testing, remove
